% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_motif_counts.R
\name{load_motif_counts}
\alias{load_motif_counts}
\title{Load motif counts from regulatory regions of interest}
\usage{
load_motif_counts(motifcountsfile)
}
\arguments{
\item{motifcountsfile}{your motif counts file in csv form}
}
\value{
the output is simply your csv counts file loaded into R
}
\description{
Load motif counts from regulatory regions of interest
}
