% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_xgb_parameters.R
\name{tune_xgb_parameters}
\alias{tune_xgb_parameters}
\title{Tune Parameters}
\usage{
tune_xgb_parameters(training_prop, parameters, nrounds, motif_counts)
}
\arguments{
\item{training_prop}{the proportion of data that should be sampled for the training set to do cross fold validation.}

\item{parameters}{a list of parameters and corresponding values you want to investigate. This should match what the parameters are called in XGBoost.}

\item{nrounds}{how many iterations of xgboost you want to perform}

\item{motif_counts}{your motif counts data frame}
}
\value{
a data frame with the parameter values tested and some evaluation metrics
}
\description{
Tune Parameters
}
